Progress dialog max for generating now set.
Apparently forgot to actually set the maximum, but the progress still
seemed to work properly?

Changed PhotomosaicGeneratorBase progress from size_t to int as
QProgressDialog uses int.

+++ src/logic/photomosaicgeneratorbase.cpp
@@ -184,7 +184,7 @@ cv::Mat PhotomosaicGeneratorBase::buildPhotomosaic(const cv::Scalar &t_backgroun
 }
 
 //Returns maximum progress
-size_t PhotomosaicGeneratorBase::getMaxProgress()
+int PhotomosaicGeneratorBase::getMaxProgress()
 {
     return std::pow(4, m_bestFits.size() - 1) * (m_bestFits.size()) *
            m_bestFits.at(0).at(0).size() * m_bestFits.at(0).size();
